Refactored responses, fixed issue with remote process call on ReviewWebApi

+++ eShop.ProductWebApi/Repositories/ProductsRepository.cs
@@ -4,17 +4,24 @@ using Azure;
 using eShop.Domain.DTOs.Requests;
 using eShop.Domain.DTOs.Responses;
 using eShop.Domain.Enums;
+using eShop.Domain.Exceptions;
+using eShop.Domain.Messages;
 using eShop.ProductWebApi.Exceptions;
 using LanguageExt;
+using MassTransit;
 using Unit = LanguageExt.Unit;
 
 namespace eShop.ProductWebApi.Repositories
 {
-    public class ProductsRepository(ProductDbContext context, IMapper mapper, ILogger<ProductsRepository> logger) : IProductRepository
+    public class ProductsRepository(ProductDbContext context,
+        IMapper mapper,
+        ILogger<ProductsRepository> logger,
+        IRequestClient<ProductDeletedMessage> requestClient) : IProductRepository
     {
         private readonly ProductDbContext context = context;
         private readonly IMapper mapper = mapper;
         private readonly ILogger<IProductRepository> logger = logger;
+        private readonly IRequestClient<ProductDeletedMessage> requestClient = requestClient;
 
         public async ValueTask<Result<IEnumerable<ProductDTO>>> GetProductsListAsync()
         {
@@ -266,10 +273,17 @@ namespace eShop.ProductWebApi.Repositories
                 var product = await context.Products.AsNoTracking().FirstOrDefaultAsync(x => x.Id == Id);
                 if (product is not null)
                 {
-                    context.Products.Remove(product);
-                    await context.SaveChangesAsync();
-                    logger.LogInformation($"Product with id: {Id} was successfully deleted.");
-                    return new(new Unit());
+                    var response = await requestClient.GetResponse<ReviewsDeletedMessage>(new ProductDeletedMessage() { Id = Id });
+                    if (response.Message.IsSucceeded)
+                    {
+                        context.Products.Remove(product);
+                        await context.SaveChangesAsync();
+                        logger.LogInformation($"Product with id: {Id} was successfully deleted.");
+                        return new(new Unit());
+                    }
+                    var failedRpcException = new FailedRpcException(response.Message.Status);
+                    logger.LogInformation($"Failed on deleting product with id: {Id} error message: {new FailedRpcException(response.Message.Status)}");
+                    return new(failedRpcException);
                 }
                 var notFoundProductException = new NotFoundProductException(Id);
                 logger.LogInformation($"Failed on deleting product with id: {Id} error message: {notFoundProductException.Message}");
